annotation_interface/metric_annotation/requirements.txt
pandas==2.2.3
python-box==7.2.0
PyYAML==6.0.2
s3fs
st-files-connection
boto3